+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * added action to list posts by month (year and month);
 * added "posts by month" widget;
 * `Photos::albums`, `Posts::categories` and `Posts::categories` widgets can now render as form or list;
+* `PostsTags::popular` widget can now render as cloud, form or list;
 * added common templates for all admin views;
 * fixed titles for some admin templates;
 * you can set a custom class for widgets.